pyproject.toml
[tool.poetry]
name = "django-rest-api"
version = "0.1.0"
description = ""
authors = ["Eugene Monnier <eugene.monnier@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
